refactor client: add service and unify competition forms
