+++ schiessmeister-client/src/components/dialogs/PlaceholderDialog.jsx
@@ -1,0 +1,9 @@
+const PlaceholderDialog = ({ message }) => {
+  return (
+    <div className="dialog">
+      <p>{message} (kommt später)</p>
+    </div>
+  );
+};
+
+export default PlaceholderDialog;
